refactoring and added envelope
