+++ examples/agent-template/agent/__init__.py
@@ -1,0 +1,1 @@
+# OpenEAGO agent-template package
